# Improve `manybodyoperator` performance once more