@aryavorskiy 2023-09-18
Done. The best way to support the vector API is implementing the `AbstractVector` interface :)
I removed the `OccupationsIterator` interface completely - one can extend this functionality by subtyping `AbstractVector` directly.
The `ManyBodyBasis` also accepts any type of containers now.