@Krastanov 2023-09-18
I do not think that `Occupations{T} <: AbstractVector{Vector{T}}` actually helps here. Subtyping in julia does not guarantee that any particular interface is being implemented or followed. I think this is an object-oriented idiom that does not really apply to Julia. Unless you are using these for something, may we just do `struct Occupations{T}` without any supertypes? Actually, if `ManyBodyBasis` can take any container, what is the reason to create `Occupations`?

Naturally, I might be wrong about these suggestions of mine, thus please run me through your reasoning to make sure I am not missing something.

Implementing formal interfaces in Julia is itself a very interesting topic with many threads on Discourse and experiments in Github repos, but no proper solution yet.