@aryavorskiy 2023-09-25
I'm sorry I disappeared for a while :)
Implementing the AbstractVector interface may be useful for non-trivial use cases like applying `map` or `find` to an `Occupations` object.
As for the `Occupations` structure itself, its main and only feature is that it wraps a **sorted** array which makes searching much faster. Is's probably better to call it in some other way (maybe `SortedContainer`?), but I think the structure itself should be kept as it is now.
I can attach some benchmarks to prove there is a significant difference in performance