@Krastanov 2023-09-25
No worries!

I agree that implementing the AbstractVector interface would be useful, but subclassing AbstractVector does not contribute to that goal, unless there are some specific methods that are already defined for AbstractVector and that rely on methods defined for Occupations.

Are you saying that your definitions of `size` and `getindex` are enough for "deriving" `map`, `find`, `length`, `collect`, etc, thanks to subtyping AbstractVector? If that is the case, I would agree with the subtyping (but I would avoid calling it an interface as I understand to have a different more formal meaning).

If these indeed work now, could you add a few "interface" unit tests? Checking that getting an index, getting a length, `collect`ing, etc all work? `collect` is probably the main one as it would be the way to fix broken user code (if it turns out that we actually broke something)