@aryavorskiy 2023-09-25
Sure, I'll rename `Occupations` to `SortedVector` and add some unit tests. The old name feels kinda bad because there is no functionality designed specially for occupation vectors.

Speaking of interfaces, I used [https://docs.julialang.org/en/v1/manual/interfaces/](this manual page) which says that redefining `size` and `getindex` along with setting `IndexStyle` to `Linear()` implements the default read-only vector semantics.