@aryavorskiy 2023-09-27
Done. 
By the way, this PR also seems to cover the issue #147 - I added a test to ensure.